+++ packages/devtools_app/test/code_size_screen_test.dart
@@ -78,6 +78,7 @@ void main() {
         codeSizeController: codeSizeController,
       );
 
+      expect(find.byType(SnapshotView), findsOneWidget);
       expect(find.byKey(CodeSizeBodyState.treemapKey), findsOneWidget);
 
       expect(
@@ -108,6 +109,7 @@ void main() {
 
         await tester.pumpAndSettle();
 
+        expect(find.byType(DiffView), findsOneWidget);
         expect(find.byKey(CodeSizeBodyState.treemapKey), findsOneWidget);
 
         expect(
@@ -116,7 +118,13 @@ void main() {
         );
 
         // Assumes the treemap is built with treemap_test_data_v8_new.json and treemap_test_data_v8_old.json
-        expect(find.text('Root [+1.5 MB]'), findsOneWidget);
+        const text = 'package:pointycastle [+465.8 KB]';
+        expect(find.text(text), findsOneWidget);
+        await tester.tap(find.text(text));
+        await tester.pumpAndSettle();
+
+        expect(find.text('ecc\n[+129.1 KB]'), findsOneWidget);
+        expect(find.text('dart:core'), findsNothing);
 
         expect(find.byType(CodeSizeSnapshotTable), findsNothing);
         expect(find.byType(CodeSizeDiffTable), findsOneWidget);