+++ tool/build_release.sh
@@ -12,7 +12,10 @@ rm -rf build
 rm -rf ../devtools/build
 flutter pub get
 
-flutter build web --dart-define=FLUTTER_WEB_USE_SKIA=true --no-tree-shake-icons
+# Build a profile build rather than a release build to avoid minification
+# as code size doesn't matter very much for us as minification makes some
+# crashes harder to debug. For example, https://github.com/flutter/devtools/issues/2125
+flutter build web --profile --dart-define=FLUTTER_WEB_USE_SKIA=true --no-tree-shake-icons
 mv build/web ../devtools/build
 
 popd