+++ packages/devtools/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.8.1-dev.1
+* Add search functionality to Timeline flame chart [#2164](https://github.com/flutter/devtools/pull/2164)
 * Add tooltips to list items in the Debugger Libraries pane [#2167](https://github.com/flutter/devtools/pull/2167)
 
 ## 0.8.0+1